Implement the fitting of PC response by spliting the real and imaginary data constructed with phase difference data and magnitude at venc=0 data. Real and imaginary are split in different files that 3dDeconvolve sees as different runs. They therefore each get their own regressors. The idea is to take the real and imaginary response regressors to construct a complex-valued response.
Overall, this computes, but there is a problem with the baseline regressors. They are independent for the real and imaginary virtual runs. This means the origin is denatured.

One should either (1) figure out proper basline model or (2) simply discard magnitude and remove any baseline model to focus only on phase. The latter should less powerfully detect activation but more precisely estimate phase responses.